Moved the metrics module to the stocks package
