+++ app/stocks/metrics.py
